优化构建过程。

+++ webpack.config.js
@@ -1,21 +1,28 @@
 var join = require("path").join;
 var ExtractTextPlugin = require("extract-text-webpack-plugin");
 var webpack = require('webpack');
-var configs = {// 基于不同模式，区分配置
+/*
+ * 基于不同模式，区分配置
+ * */
+var configs = {
     "development": {
-        "path": './public',
+        "path": './dist',
         "publicPath": '/s/'
     },
     "test": {
-        "path": './public',
+        "path": './dist',
         "publicPath": '/s/'
     },
     "production": {
-        "path": './public',
+        "path": './dist',
         "publicPath": '/s/'
     }
 };
 var cfg = configs.development;
+/*
+ * 构建之前是否清除之前构建的内容。
+ * */
+var clean = true;
 /*
  * 获取自定义参数
  * 自定义参数约定：webpack --cfg.path=./public --cfg.runmod=dev
@@ -25,18 +32,17 @@ var cfg = configs.development;
  *
  var argv = require('argv');
  var args = argv.option({
-     name: 'option',
-     short: 'o',
-     type: 'string',
-     description: 'Defines an option for your script',
-     example: "'script --opiton=value' or 'script -o value'"
+ name: 'option',
+ short: 'o',
+ type: 'string',
+ description: 'Defines an option for your script',
+ example: "'script --opiton=value' or 'script -o value'"
  }).run();
 
  console.dir(args.options);
  然后
  $ node app.js -o aaa
  { option: 'aaa' }
- *
  * */
 var arguments = process.argv.splice(2);
 if (arguments) {
@@ -50,43 +56,66 @@ if (arguments) {
         if (cfgs[0] == 'runmod' && configs[cfgs[1]]) {
             cfg = configs[cfgs[1]];
         }
+        if (cfgs[0] == 'clean' && cfgs[1] == 'false') {
+            clean = false;
+        }
     }
 }
 /*
-* 定义entry
-* 如果项目结构命名有良好的约定，是否考虑使用代码自动生成entry？
-*
-* */
-
+ * 定义entry
+ * 如果项目结构命名有良好的约定，是否考虑使用代码自动生成entry？
+ * */
 var _entry = {
     //"index": ["./src/home/home.jsx", "./src/home/home-content.jsx"],//会合并成一个index.js
     "index": './src/entry/index.jsx'
 };
 
-//构建之前，先清除之前构建的内容
-var child_process = require('child_process');
-child_process.exec('rm -rf ' + join(__dirname, cfg.path),
-    function (error, stdout, stderr) {
-        if (error !== null) {
-            console.log('exec error: ' + error);
-        }
-    });
+/*
+ * 构建之前，先清除之前构建的内容
+ * 注：这样有个问题，执行 webpack-dev-server 的时候会删除对应的构建内容，并且没有再生成。
+ * */
+if (clean) {
+    var child_process = require('child_process');
+    child_process.exec('rm -rf ' + join(__dirname, cfg.path),
+        function (error, stdout, stderr) {
+            if (error !== null) {
+                console.log('exec error: ' + error);
+            }
+        });
+}
 
+/*
+ * babel参数
+ * */
 var babelQuery = {
     presets: ['es2015', 'react', 'stage-0'],
     plugins: ['add-module-exports', 'typecheck']
 };
 
+/*
+ * webpack配置
+ * */
 module.exports = {
-    resolve: {//加入这个，import js或者jsx文件时，可以忽略后缀名
+    /*
+     * import js或者jsx文件时，可以忽略后缀名
+     * */
+    resolve: {
         modulesDirectories: ['node_modules', (0, join)(__dirname, '../node_modules')],
-        extensions: ['', '.js', '.jsx']
+        extensions: ['', '.js', '.jsx'],
     },
-
     resolveLoader: {
         modulesDirectories: ['node_modules', (0, join)(__dirname, '../node_modules')]
     },
-    entry:_entry,
+    /*
+     * 入口文件配置
+     * */
+    entry: _entry,
+    /*
+     * 输出配置
+     * path：构建之后的文件存放目录
+     * publicPath：js或css等文件，浏览器访问时路径
+     * filename：构建之后的文件名
+     * */
     output: {
         path: join(__dirname, cfg.path),
         publicPath: cfg.publicPath,
@@ -133,8 +162,15 @@ module.exports = {
         ]
     },
     plugins: [
+        /*
+         * 公共文件配置
+         * */
         new webpack.optimize.CommonsChunkPlugin('common', 'common.js'),
-        new ExtractTextPlugin("[name].css", {// css单独打包成一个css文件，比如entry.js引入了多个less，最终会都打到一个home.css中。
+        /*
+         * css单独打包成一个css文件
+         * 比如entry.js引入了多个less，最终会都打到一个xxx.css中。
+         * */
+        new ExtractTextPlugin("[name].css", {
             disable: false,
             allChunks: true
         }),
